requirements.txt
pdfplumber
python-docx
textract
requests
python-dotenv
rich
